Cheffing (#30)
* stage 1

* routines now add but no display?

* attempted routes

* savepoint

* save

* Creat Route works just to style

* TBC

* Create works

* Handing over to myself

* Cheffing again

* fixed some more stuff

* add-button not wokring

* add button works

* view button fix

* got routine to show

* edited routines

* delete yesssss

* everything besides delete exercise and unwanted refreshes

* delete button for specific exercises work now

* just need to fix edit name and css

* DONE, just css now

* Finito

* jonathan name fix

---------

+++ App/controllers/routine.py
@@ -1,13 +1,14 @@
 from App.models import Routine
 from App.database import db
+from .workoutroutine import get_all_workouts_in_routines
 
    return routine
 
-def create_routine(name, user_id):
-    routine = Routine(name = name, user_id=user_id)
-    db.session.add(routine)
-    db.session.commit()
-    return routine
+# def create_routine2(name, description, user_id):
+#     routine = Routine(name=name, description=description, user_id=user_id)
+#     db.session.add(routine)
+#     db.session.commit()
+#     return routine
 
 def get_all_routines():
     return Routine.query.all()
@@ -16,27 +17,36 @@ def get_routine_by_name(name):
     return Routine.query.filter_by(name=name).first()
 
 def get_routine_by_id(id):
-    return Routine.query.get(id)
+    routine = Routine.query.filter_by(routine_id=id).first()
+    return routine
 
 def get_user_routines(user_id):
     return Routine.query.filter_by(user_id=user_id).all()
 
-def update_routine(id, name=None, description=None):
+def get_all_routines_json():
+    routines = get_all_routines()
+
+    if not routines:
+        return[]
+
+    return [routine.get_json() for routine in routines]
+
+def update_routine(id, name=None):
     routine = Routine.query.get(id)
     if name is not None:
         routine.name = name
-    if description is not None:
-        routine.description = description
     db.session.commit()
     return routine
 
-def add_workout_to_routine(routine_id, workout_id):
-    routine = Routine.query.get(routine_id)
-    workout = Workout.query.get(workout_id)
-    routine.workouts.append(workout)
-    db.session.commit()
 
 def delete_routine(id):
-    routine = Routine.query.get(id)
-    db.session.delete(routine)
-    db.session.commit()
+    routine = Routine.query.filter_by(routine_id=id).first()
+    workouts = get_all_workouts_in_routines(id)
+    if routine:
+        db.session.delete(routine)
+        for workout in workouts: 
+            db.session.delete(workout)
+        db.session.commit()
+        return routine
+    else:
+        return None
